Refactored the package layout in the ripple-security to make more sense
