+++ ripple-security/src/main/java/org/rippleosi/login/api/rest/LoginController.java
@@ -13,7 +13,7 @@
  *    See the License for the specific language governing permissions and
  *    limitations under the License.
  */
-package org.rippleosi.api.login.rest;
+package org.rippleosi.login.api.rest;
 
 import javax.annotation.PostConstruct;
 import javax.servlet.http.HttpServletRequest;
