+++ ripple-security/src/main/java/org/rippleosi/logout/api/LogoutController.java
@@ -1,4 +1,4 @@
-package org.rippleosi.api.logout.rest;
+package org.rippleosi.logout.api;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
